Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-css-xml-attributeType-expected.txt


FAIL Tests that XML and CSS attributeTypes can be switched between. assert_approx_equals: expected 100 +/- 1 but got 300
